+++ moesland-app/src/utilities/HelperFunctions.js
