+++ moesland-app/src/views/CalendarView.jsx
@@ -5,7 +5,7 @@ import { LocaleConfig } from 'react-native-calendars';
 import styles from '../styles/views/CalendarViewStyles';
 import testIDs from '../constants/calendarTestIDs';
 import fetchEvents from '../services/EventApi';
-import { timeToString } from '../lib/Utilities/HelperFunctions';
+import { timeToString } from '../Utilities/HelperFunctions';
 
 
 export default class AgendaScreen extends PureComponent {
